test_site/mysite/requirements.txt
asgiref==3.3.4
Django==3.2.3
django-cart==1.0.4
django-debug-toolbar==3.2.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
djangorestframework==3.12.4
Pillow==8.2.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
